Update patch set 1
Patch Set 1:

The tethering offload HAL is entirely separate from this HAL. In the case of the tethering offload HAL, the client is the framework (system_server) and the implementation is in vendor code. In the case of this HAL, the client will be vendor code and the implementation will be in netd.

Patch-set: 1
Reviewer: Gerrit User 1000835 <1000835@85c56323-6fa9-3386-8a01-6480fb634889>