Update patch set 1
Patch Set 1:

> Adding Tom Cherry in case he can think of a better way to do this in init.

There can be an `onrestart` option added to netd.  It could restart a specifically named class and everything that needs to be restarted along with netd can be added there.

But... I sort of figure if there's going to be a HAL in the future anyway and once the HAL is in place the death notifications are free, then maybe the HAL is fine?

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>