Update patch set 1
Patch Set 1:

Discussed with Niraj and the plan is to add two operations (create OEM networks and destroy OEM networks) to the initial release of this HAL. These operations are currently being performed by shelling out to ndc, but that imposes performance penalties.

The plan is still that access from proprietary vendor code to the networking stack will be mediated to this HAL, so we expect more revisions in the future.

Patch-set: 1
Reviewer: Gerrit User 1000835 <1000835@85c56323-6fa9-3386-8a01-6480fb634889>