Make the VTS tests for the INetd HAL 1.0 more realistic.
The VTS test for the netd HAL creates an OEM network, but it
doesn't check whether the network exists. Fix that by using
the native multinetwork API calls to select the network.
Also, check that when an OEM network is created, there are no IP
rules that select for its fwmark. There should be none because
a network that has just been created has no interfaces.

Put utility code into a new VtsHalNetNetdTestUtils static library
to avoid code duplication when we add new tests for future HALversions.

Bug: 73032258
Test: VtsHalNetNetdV1_0TargetTest passes on walleye
Change-Id: I345b9b27d7c2c8bdd6ce6d60430490dac8c3a989