Update patch set 2
Patch Set 2:

Actually, relying on the destructors may introduce a slight lag as well (as they are not guaranteed to be sent immediately). This change probably is for the better, but I like relying on the destructors because the slight delay shouldn't matter in this case, and having exactly one way to do something is nice. Also, I think 'myWakeLock.clear()' is pretty readable.

Patch-set: 2
Reviewer: Gerrit User 1120458 <1120458@85c56323-6fa9-3386-8a01-6480fb634889>