Passing partial vs full wake lock info for ARC++
We pass an enum to acquireWakeLock(), symmetric to how it's done in libpower
interface.

The screen could be managed above (android) or below (ARC++) this HAL depending
on the device, so we don't mandate that full wake locks keep the screen alive.
Only requirement is that all wake locks block system suspend.

Bug: 115947180
Test: SystemSuspendV1_0UnitTest
Change-Id: I92c65371852ffb2f71f5d529b7020d256cb1ba76