Update patch set 1
Patch Set 1:

Since this interface is being exposed to PowerManager on the system partition, we should use AIDL instead of HIDL to express the interface definition. Otherwise, we are actually exposing the ABI to vendors. This will ensure you can continue to modify the API in the future as needed (vs. having to deal w/ strong versioning requirements for vendor/system boundary)

Patch-set: 1
Reviewer: Gerrit User 1120458 <1120458@85c56323-6fa9-3386-8a01-6480fb634889>