Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-2
> 
> There must not be any other way to suspend the device. How does the spurious suspension happen then?

Tri, I'm not sure what you're asking.

My thinking is that if SystemSuspend is instantiated with mUseSuspendCounter false,
then autosuspendThread should not be used. By adding the extra condition to
autosuspendThread's wait() call, we ensure that the wait() never completes, and so
the thread does not take action.

I suppose that with mUseSuspendCounter false, the only time the wait() will check its
condition is in a spurious wake.

Patch-set: 1
Reviewer: Gerrit User 1074960 <1074960@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0