Move suspend AIDL into suspend.
This AIDL interfaces was added before AIDL interfaces could be used as
a hardare interface, and by chance, it is located in the location that a
hardare interface AIDL is expected.

Since this can be regarded as an implementation detail of the sole
platform service implementation in 'default', I've moved it there and
renamed the folder for further emphasis.

Fixes: 152769355
Test: m nothing (sanity)
Change-Id: Iff685ba3790cb0ee85767108387e0ac2fc3e1f1d