system_suspend: fix null pointer crash when multiple clients binderDied
When multiple registered entities triggered binderDead, two scenarios
can cause the crash:
1. When handling the binderDead with a client registered for wakelock
but not wakeup notification. If the callback list for wakeup is empty,
SystemSuspend will try erase an end() iterator.

2. After erasing the last callback from a wakelock named vector, the
pair of wakelock name and the empty vector will be removed from the
map. If such pair is positioned at last of the map, further increment
of the iterator by the loop operation will cause crash.

Test: atest SystemSuspendV1_0UnitTest
Bug: 173174189
Change-Id: I183ea631cd3289902197f86348b6216e98c850bc