Adding new error to ResponseCode
This change adds OUT_OF_KEYS to the ReponseCode.aidl for keystore2. This
error is intended to be thrown to the keystore2 SPI when the attestation
key pool for a given KM instance is out of server signed attestation
keys to assign to apps. When the SPI sees this, it can attempt to poke
the RemoteProvisioner app to immediately request new keys to be signed
and then retry key generation.

Test: the aidl builds

Change-Id: I500e22d5a05b431f78160523c379a059a681233d