system_suspend: allow SystemSuspend to re-open the fd if found stale
Based on the BR and debug message in system_suspend service it is
evedent that if kernel fails to populate any wakeup reason in
/sys/kernel/wakeup_reasons/last_resume_reason, i.e. if the file is
found empty when system_suspend tries to read, then the fd is unable
to read the file anymore, and all the subsequent attempts to read the
wakeup_reason returns as "unknown" which causes flurry of "unknwon"
wakeup reasons.

Kernel populates the sysfs node upon read() call to the sysfs node,
and hence it is not possible to know ahead of time if and when kernel
will leave the file empty with no wakeup reason.

The way we can recover from this condition is by re-opening the fd upon
failure (or when read call returns with 0 bytes), but the SystemSuspend
currently does not have any knowledge about the file path. This change
uses a file factory method and main will create and share the objects
holding file info with SystemSuspend rather than just sharing file
descriptors.

Bug: 192813633
Test: Since the repo rate is too low, a kernel change was made in
wakeup_reason.c to periodically setting empty wakeup reason.

Signed-off-by: Manish Varma <varmam@google.com>
Change-Id: I2769b739aba801ac89c88f4811877514df062cb2