suspend: Reenable check system server liveness before suspend
Taking the kernel wakelock on zygote restart can happen after suspend
is initiated. The new wakelock will not abort an already in-progress
suspend operation.

Writing 'on' to /sys/power/state is a no-op and doesn't abort an
in-progress suspend.

Reenable the previos approach of checking that system_server is alive
before attempting a suspend.

Bug: 304183174
Test: atest SystemSuspendV1_0UnitTest
Change-Id: Ia13e4aaf6c73fde9d358496223808e75bcc48955