suspend: Add test for multiple wakelock releases
Add a test case to verify that multiple releases of a wakelock after
the initial release have no effect on the `isActive`, `activeCount`,
and `activeTime` fields in the corresponding `WakeLockInfo` object.

`Wakelock.release()` AIDL interface is marked `oneway` which means
multiple calls are guarenteed to have no effect. This test case is
added just to safeguard against future changes.

Test: atest SystemSuspendV1_0UnitTest
Bug: 356957253
Change-Id: Iecec72958110d73f67b55745da1256ee2e34fce6